+++ crates/tako/src/internal/scheduler/batches.rs
@@ -1,5 +1,6 @@
 use crate::Priority;
 use crate::internal::server::core::{Core, CoreSplitMut};
+use crate::internal::server::worker::Worker;
 use crate::resources::ResourceRqId;
 use std::cmp::Ordering;
 use std::time::Instant;
@@ -30,7 +31,11 @@ impl TaskBatch {
     }
 }
 
-pub(crate) fn create_task_batches(core: &mut Core, now: Instant) -> Vec<TaskBatch> {
+pub(crate) fn create_task_batches(
+    core: &mut Core,
+    now: Instant,
+    custom_workers: Option<&[Worker]>,
+) -> Vec<TaskBatch> {
     let CoreSplitMut {
         task_map,
         worker_map,
@@ -67,8 +72,9 @@ pub(crate) fn create_task_batches(core: &mut Core, now: Instant) -> Vec<TaskBatc
                     .sum::<u32>();
                 n_frees / n_nodes
             } else {
-                worker_map
-                    .get_workers()
+                custom_workers
+                    .map(|ws| itertools::Either::Right(ws.iter()))
+                    .unwrap_or(itertools::Either::Left(worker_map.get_workers()))
                     .map(|w| {
                         w.sn_assignment()
                             .map(|a| a.free_resources.task_max_count(&rqv))